+++ .editorconfig
@@ -9,6 +9,10 @@ insert_final_newline = true
 indent_style = space
 indent_size = 4
 
+[zmarkdown/**/*.js]
+indent_style = space
+indent_size = 2
+
 [Makefile]
 indent_style = tab
