+++ package.json
@@ -44,7 +44,7 @@
     "jquery": "3.2.1",
     "normalize.css": "7.0.0",
     "pm2": "^2.6.1",
-    "zmarkdown": "^1.0.0"
+    "zmarkdown": "^1.0.1"
   },
   "devDependencies": {
     "gulp-jshint": "2.0.4",
